tools: Move data structures to data_structures.py
