+++ devito/tools/utils.py
@@ -1,23 +1,18 @@
 from abc import ABC
 import os
 import ctypes
-from collections import Callable, Iterable, OrderedDict, Mapping
+from collections import Iterable, OrderedDict
 from functools import reduce
 from itertools import chain, combinations, product, zip_longest
 from operator import attrgetter, mul
 
 import numpy as np
 
-from multidict import MultiDict
-
-from devito.logger import error
-
 __all__ = ['prod', 'as_tuple', 'is_integer', 'generator', 'grouper', 'split',
            'roundm', 'powerset', 'invert', 'flatten', 'single_or', 'filter_ordered',
            'filter_sorted', 'numpy_to_ctypes',
-           'ctypes_to_C', 'ctypes_pointer', 'pprint', 'DefaultOrderedDict',
-           'change_directory', 'Bunch',
-           'EnrichedTuple', 'ReducerMap', 'Tag', 'sweep']
+           'ctypes_to_C', 'ctypes_pointer', 'pprint',
+           'change_directory', 'Tag', 'sweep']
 
 
 def prod(iterable):
@@ -180,41 +175,6 @@ def pprint(node, verbose=True):
     print(printAST(node, verbose))
 
 
-class DefaultOrderedDict(OrderedDict):
-    # Source: http://stackoverflow.com/a/6190500/562769
-    def __init__(self, default_factory=None, *a, **kw):
-        if (default_factory is not None and
-           not isinstance(default_factory, Callable)):
-            raise TypeError('first argument must be callable')
-        OrderedDict.__init__(self, *a, **kw)
-        self.default_factory = default_factory
-
-    def __getitem__(self, key):
-        try:
-            return OrderedDict.__getitem__(self, key)
-        except KeyError:
-            return self.__missing__(key)
-
-    def __missing__(self, key):
-        if self.default_factory is None:
-            raise KeyError(key)
-        self[key] = value = self.default_factory()
-        return value
-
-    def __reduce__(self):
-        if self.default_factory is None:
-            args = tuple()
-        else:
-            args = self.default_factory,
-        return type(self), args, None, None, self.items()
-
-    def copy(self):
-        return self.__copy__()
-
-    def __copy__(self):
-        return type(self)(self.default_factory, self)
-
-
 class change_directory(object):
     """
     Context manager for changing the current working directory.
@@ -246,103 +206,6 @@ def sweep(parameters, keys=None):
                     for v in sweep_values]
     for vals in product(*sweep_values):
         yield dict(zip(keys, vals))
-
-
-class Bunch(object):
-    """
-    Bind together an arbitrary number of generic items. This is a mutable
-    alternative to a ``namedtuple``.
-
-    From: ::
-
-        http://code.activestate.com/recipes/52308-the-simple-but-handy-collector-of\
-        -a-bunch-of-named/?in=user-97991
-    """
-    def __init__(self, **kwargs):
-        self.__dict__.update(kwargs)
-
-
-class EnrichedTuple(tuple):
-    """
-    A tuple with an arbitrary number of additional attributes.
-    """
-    def __new__(cls, *items, getters=None, **kwargs):
-        obj = super(EnrichedTuple, cls).__new__(cls, items)
-        obj.__dict__.update(kwargs)
-        obj._getters = dict(zip(getters or [], items))
-        return obj
-
-    def __getitem__(self, key):
-        if isinstance(key, int):
-            return super(EnrichedTuple, self).__getitem__(key)
-        else:
-            return self._getters[key]
-
-
-class ReducerMap(MultiDict):
-    """
-    Specialised :class:`MultiDict` object that maps a single key to a
-    list of potential values and provides a reduction method for
-    retrieval.
-    """
-
-    def update(self, values):
-        """
-        Update internal mapping with standard dictionary semantics.
-        """
-        if isinstance(values, Mapping):
-            self.extend(values)
-        elif isinstance(values, Iterable) and not isinstance(values, str):
-            for v in values:
-                self.extend(v)
-        else:
-            self.extend(values)
-
-    def unique(self, key):
-        """
-        Returns a unique value for a given key, if such a value
-        exists, and raises a ``ValueError`` if it does not.
-
-        :param key: Key for which to retrieve a unique value
-        """
-        candidates = self.getall(key)
-
-        def compare_to_first(v):
-            first = candidates[0]
-            if isinstance(first, np.ndarray) or isinstance(v, np.ndarray):
-                return (first == v).all()
-            else:
-                return first == v
-
-        if len(candidates) == 1:
-            return candidates[0]
-        elif all(map(compare_to_first, candidates)):
-            return candidates[0]
-        else:
-            error("Unable to find unique value for key %s, candidates: %s" %
-                  (key, candidates))
-            raise ValueError('Inconsistent values for key reduction')
-
-    def reduce(self, key, op=None):
-        """
-        Returns a reduction of all candidate values for a given key.
-
-        :param key: Key for which to retrieve candidate values
-        :param op: Operator for reduction among candidate values.
-                   If not provided, a unique value will be returned,
-                   or a ``ValueError`` raised if no unique value exists.
-        """
-        if op is None:
-            # Return a unique value if it exists
-            return self.unique(key)
-        else:
-            return reduce(op, self.getall(key))
-
-    def reduce_all(self):
-        """
-        Returns a dictionary with reduced/unique values for all keys.
-        """
-        return {k: self.reduce(key=k) for k in self}
 
 
 class Tag(ABC):
